apps/sandbox_cli/src/main.py
"""
E2B Sandbox CLI - Main entry point.

A comprehensive CLI for managing E2B sandboxes and performing operations.
"""

import click
from pathlib import Path
from dotenv import load_dotenv
from rich.console import Console

# Load environment variables
root_dir = Path(__file__).parent.parent.parent.parent
load_dotenv(root_dir / ".env")

# Import command groups
from .commands.sandbox import sandbox
from .commands.files import files
from .commands.exec import exec

console = Console()


@click.group()
@click.version_option(version="0.1.0")
def cli():
    """
    E2B Sandbox CLI - Control sandboxes from the command line.

    This CLI provides comprehensive sandbox management capabilities:
    - Create, connect to, and manage sandboxes (sandbox)
    - Perform file operations with SDK APIs (files)
    - Execute any command with full control (exec)

    Most commands require a SANDBOX_ID. You can get one by:
    1. Creating a new sandbox: sbx init
    2. Or: sbx sandbox create

    Tip: Export your sandbox ID for easier use:
      export SANDBOX_ID=<your-sandbox-id>
      sbx files ls $SANDBOX_ID /
      sbx exec $SANDBOX_ID "python --version"
    """
    pass


# Add command groups
cli.add_command(sandbox)
cli.add_command(files)
cli.add_command(exec)


# Add an init command for quick sandbox setup
@cli.command()
@click.option(
    "--template",
    "-t",
    default=None,
    help="Sandbox template name or ID",
)
@click.option(
    "--timeout", default=600, help="Sandbox timeout in seconds (default: 10 minutes)"
)
@click.option("--env", "-e", multiple=True, help="Environment variables (KEY=VALUE)")
def init(template, timeout, env):
    """
    Initialize a new sandbox and display the ID.

    This is a convenience command that creates a sandbox and
    saves the ID to a local .sandbox_id file.

    Templates:
        You can create a sandbox from a custom template with pre-installed tools.
        Use --template to specify a template name.

    Examples:
        # Create with default base template
        sbx init

        # Create with custom template and environment
        sbx init --template agent-sandbox-dev-node22 --env API_KEY=secret --timeout 3600
    """
    try:
        from .modules import sandbox as sbx_module

        console.print("[yellow]Initializing new sandbox...[/yellow]")
        if template:
            console.print(f"[dim]Template: {template}[/dim]")

        # Parse env vars
        envs = {}
        for e in env:
            if "=" in e:
                key, value = e.split("=", 1)
                envs[key] = value

        sbx = sbx_module.create_sandbox(
            template=template, timeout=timeout, envs=envs if envs else None
        )

        console.print(f"\n[green]✓ Sandbox created successfully![/green]")
        console.print(f"\n[cyan]Sandbox ID:[/cyan] {sbx.sandbox_id}")
        if template:
            console.print(f"[dim]Template: {template}[/dim]")

    except Exception as e:
        console.print(f"[red]✗ Error: {e}[/red]")
        raise click.Abort()


if __name__ == "__main__":
    cli()
